requirements.txt
Requirements:
pillow
pygame
pystray